+++ tests/testthat/test-seriesname.R
@@ -1,0 +1,7 @@
+context("Checking seriesname")
+
+test_that("seriesname ...",{
+
+
+})
+